requirements.txt
dj-database-url==0.5.0
Django==2.2.5
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.10.3
gunicorn==19.9.0
psycopg2==2.8.3
pytz==2019.2
PyYAML==5.1.2
sqlparse==0.3.0
whitenoise==4.1.3
requests==2.22.0